finished product routes
